requirements.txt
mcp>=1.0.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b
requests>=2.31.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
scipy>=1.11.0
sqlalchemy>=2.0.0
httpx>=0.27.0
